+++ packages/smithers/flows/crypto/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @smthrs/crypto
 
+## Unreleased
+
+### Fixed
+
+- Share live function identity metadata across compatible bundles without reading entropy at module initialization. ([#1991](https://github.com/smithersai/smithers/issues/1991)).
+
 ## [1.0.0-rc.0] - 2026-08-31
 
 ### Added